requirements.txt
django
jsonschema
djangorestframework
djangorestframework-jsonp
django_object_hooks
isodate